consolidate asio io_context benchmarks to use coroutines only
Remove the separate lambda-based benchmark and keep only the coroutine
version for fair comparison with corosio. Both libraries now benchmark
coroutine dispatch rather than comparing coroutines against callbacks.
